refactor: add customized error for missing error
