+++ src/presentation/errors/missing-param-error.ts
@@ -1,0 +1,6 @@
+export class MissingParamError extends Error {
+  constructor(paramName: string) {
+    super(`Missing param: ${paramName}`);
+    this.name = "MissingParamError";
+  }
+}
