+++ tests/presentation/controllers/signup-controller.spec.ts
@@ -1,0 +1,31 @@
+import { SignUpController } from "@/presentation/controllers/signup-controller";
+import { MissingParamError } from "@/presentation/errors/missing-param-error";
+
+describe("SignUpController", () => {
+  test("Should return 400 if no name is provided", () => {
+    const sut = new SignUpController();
+    const httpRequest = {
+      body: {
+        email: "any_email@mail.com",
+        password: "any_password",
+        passwordConfirmation: "any_password",
+      },
+    };
+    const httpResponse = sut.handle(httpRequest);
+    expect(httpResponse.statusCode).toBe(400);
+    expect(httpResponse.body).toEqual(new MissingParamError("name"));
+  });
+  test("Should return 400 if no email is provided", () => {
+    const sut = new SignUpController();
+    const httpRequest = {
+      body: {
+        name: "any_name",
+        password: "any_password",
+        passwordConfirmation: "any_password",
+      },
+    };
+    const httpResponse = sut.handle(httpRequest);
+    expect(httpResponse.statusCode).toBe(400);
+    expect(httpResponse.body).toEqual(new MissingParamError("email"));
+  });
+});
